backend/requirements.txt
fastapi
uvicorn
pydantic
google-generativeai
python-dotenv
pymupdf
python-multipart
psycopg2-binary
supabase
sqlalchemy
